[IMP] ePoS: vulgarise the SSL ePos issue.
SSL/HTTPS topic is complicated for most of
Odoo customers as it is quite technical.

This PR should help them guide them to better
understand the issue and how to fix it themselves.

Support can't be provided to each device, browsers and OS.
But we did add some guides regarding the more
"popular" ones and some "keyword" to search
online for the others.

Signed-off-by: Loan (LSE) <lse@odoo.com>